visualization/pyproject.toml
[tool.poetry]
name = "annotation-stat"
version = "0.1.0"
description = ""
authors = ["t1910401 <t1910401@gl.cc.uec.ac.jp>"]
readme = "README.md"
packages = [{include = "annotation_stat"}, {include = "src/dataset"}]

[tool.poetry.dependencies]
python = "^3.10,<3.12"
seaborn = "^0.12.1"
pandas = "^1.5.2"
numpy = "^1.23.5"
unidic = "^1.1.0"
mecab-python3 = "^1.0.5"
unidic-lite = "^1.0.8"
wordcloud = "^1.8.2.2"
japanize-matplotlib = "^1.1.3"
jaconv = "^0.3"
scipy = "^1.10.0"
pyyaml = "^6.0"
hydra-core = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
